Split object_cache_srv into modules and add README
