+++ rust/object-cache-srv/src/app_state.rs
@@ -1,0 +1,7 @@
+use micromegas_range_cache::range_cache::RangeCache;
+
+#[derive(Clone)]
+pub(crate) struct AppState {
+    pub(crate) cache: RangeCache,
+    pub(crate) allowed_prefix: String,
+}
