+++ rust/object-cache-srv/src/object_cache_srv.rs
@@ -2,380 +2,35 @@
 #[global_allocator]
 static ALLOC: tikv_jemallocator::Jemalloc = tikv_jemallocator::Jemalloc;
 
-use anyhow::{Context, Result, anyhow, bail};
+mod app_state;
+mod cli;
+mod handlers;
+mod validation;
+
+use anyhow::{Context, Result, anyhow};
+use app_state::AppState;
 use axum::{
     Router,
-    body::Body,
-    extract::{Path, State},
-    http::{HeaderMap, StatusCode},
+    http::StatusCode,
     middleware,
-    response::Response,
     routing::{get, post},
 };
-use bytes::{BufMut, Bytes, BytesMut};
 use clap::Parser;
+use cli::Cli;
+use handlers::{get_range_handler, head_handler, post_ranges_handler};
 use micromegas::micromegas_main;
 use micromegas::servers::shutdown::{serve_axum_with_graceful_shutdown, wait_for_sigterm};
 use micromegas_auth::api_key::{ApiKeyAuthProvider, parse_key_ring};
 use micromegas_auth::axum::auth_middleware;
 use micromegas_auth::types::AuthProvider;
 use micromegas_range_cache::foyer_backend::FoyerBackend;
-use micromegas_range_cache::range_cache::{RangeCache, RangeError};
+use micromegas_range_cache::range_cache::RangeCache;
 use micromegas_tracing::prelude::*;
 use object_store::parse_url_opts;
 use object_store::prefix::PrefixStore;
-use serde::Deserialize;
 use std::net::SocketAddr;
 use std::sync::Arc;
 use std::time::Duration;
-
-#[derive(Parser, Debug)]
-#[clap(name = "micromegas-object-cache-srv")]
-#[clap(about = "Shared object range cache service", version, author)]
-struct Cli {
-    #[clap(
-        long,
-        env = "MICROMEGAS_OBJECT_CACHE_LISTEN",
-        default_value = "0.0.0.0:8080"
-    )]
-    listen: SocketAddr,
-
-    #[clap(long, env = "MICROMEGAS_OBJECT_CACHE_ORIGIN_URI")]
-    origin_uri: String,
-
-    #[clap(long, env = "MICROMEGAS_OBJECT_CACHE_RAM_MB", default_value = "512")]
-    ram_mb: usize,
-
-    #[clap(long, env = "MICROMEGAS_OBJECT_CACHE_DISK_PATH")]
-    disk_path: String,
-
-    #[clap(long, env = "MICROMEGAS_OBJECT_CACHE_DISK_GB", default_value = "50")]
-    disk_gb: usize,
-
-    #[clap(
-        long,
-        env = "MICROMEGAS_OBJECT_CACHE_BLOCK_SIZE",
-        default_value = "1048576"
-    )]
-    block_size: u64,
-
-    #[clap(long, env = "MICROMEGAS_OBJECT_CACHE_NAMESPACE", default_value = "")]
-    namespace: String,
-
-    #[clap(long, env = "MICROMEGAS_OBJECT_CACHE_PREFIX", default_value = "")]
-    allowed_prefix: String,
-
-    #[clap(long, env = "MICROMEGAS_API_KEYS", default_value = "")]
-    api_keys: String,
-
-    /// Disable authentication (development mode only)
-    #[clap(long)]
-    disable_auth: bool,
-
-    #[clap(
-        long,
-        default_value = "25",
-        env = "MICROMEGAS_SHUTDOWN_GRACE_PERIOD_SECONDS"
-    )]
-    shutdown_grace_period_seconds: u64,
-}
-
-#[derive(Clone)]
-struct AppState {
-    cache: RangeCache,
-    allowed_prefix: String,
-}
-
-fn validate_key(key: &str, allowed_prefix: &str) -> Result<()> {
-    if key.is_empty() {
-        bail!("empty key");
-    }
-    if key.starts_with('/') {
-        bail!("key must not start with /");
-    }
-    if key.split('/').any(|seg| seg == "..") {
-        bail!("key must not contain ..");
-    }
-    if !allowed_prefix.is_empty()
-        && key != allowed_prefix
-        && !key.starts_with(&format!("{allowed_prefix}/"))
-    {
-        bail!("key {key} is outside allowed prefix {allowed_prefix}");
-    }
-    Ok(())
-}
-
-fn parse_range_header(header_value: &str, file_size: u64) -> Result<std::ops::Range<u64>> {
-    let value = header_value
-        .strip_prefix("bytes=")
-        .ok_or_else(|| anyhow!("invalid Range header: {header_value}"))?;
-    let (start_str, end_str) = value
-        .split_once('-')
-        .ok_or_else(|| anyhow!("invalid Range header format: {header_value}"))?;
-    let start: u64 = start_str.parse().with_context(|| "parsing range start")?;
-    if end_str.is_empty() {
-        // Open-ended range (`bytes=<start>-`): read from `start` to EOF. An
-        // offset exactly at EOF is a legitimate zero-length read in
-        // `object_store::GetRange::Offset` semantics, so allow `start == file_size`
-        // to yield an empty range rather than rejecting it. Note `end > file_size`
-        // (i.e. `start > file_size`) is left for the caller's OutOfBounds→416 path.
-        Ok(start..file_size)
-    } else {
-        let end = end_str
-            .parse::<u64>()
-            .with_context(|| "parsing range end")?
-            .checked_add(1)
-            .ok_or_else(|| anyhow!("range end overflow in Range header: {header_value}"))?;
-        // Reject inverted/degenerate explicit ranges (e.g. `bytes=100-50`): an
-        // empty or backwards range cannot produce a valid 206 Content-Range.
-        if start >= end {
-            bail!("invalid Range header: start {start} not before end {end}");
-        }
-        Ok(start..end)
-    }
-}
-
-fn is_not_found(e: &anyhow::Error) -> bool {
-    if let Some(os_err) = e.downcast_ref::<object_store::Error>() {
-        matches!(os_err, object_store::Error::NotFound { .. })
-    } else {
-        false
-    }
-}
-
-async fn head_handler(
-    Path(key): Path<String>,
-    State(state): State<AppState>,
-) -> Result<Response, StatusCode> {
-    if let Err(e) = validate_key(&key, &state.allowed_prefix) {
-        warn!("rejected key {key}: {e}");
-        return Err(StatusCode::BAD_REQUEST);
-    }
-    match state.cache.size(&key).await {
-        Ok(size) => {
-            let response = Response::builder()
-                .status(StatusCode::OK)
-                .header("Content-Length", size.to_string())
-                .body(Body::empty())
-                .expect("build HEAD response");
-            Ok(response)
-        }
-        Err(e) => {
-            if is_not_found(&e) {
-                Err(StatusCode::NOT_FOUND)
-            } else {
-                error!("HEAD {key}: {e:?}");
-                Err(StatusCode::INTERNAL_SERVER_ERROR)
-            }
-        }
-    }
-}
-
-async fn get_range_handler(
-    Path(key): Path<String>,
-    State(state): State<AppState>,
-    headers: HeaderMap,
-) -> Result<Response, StatusCode> {
-    if let Err(e) = validate_key(&key, &state.allowed_prefix) {
-        warn!("rejected key {key}: {e}");
-        return Err(StatusCode::BAD_REQUEST);
-    }
-
-    let file_size = match state.cache.size(&key).await {
-        Ok(s) => s,
-        Err(e) => {
-            if is_not_found(&e) {
-                return Err(StatusCode::NOT_FOUND);
-            }
-            error!("size {key}: {e:?}");
-            return Err(StatusCode::INTERNAL_SERVER_ERROR);
-        }
-    };
-
-    // A zero-byte object cannot be expressed as a satisfiable byte range, and
-    // `Content-Range: bytes 0-0/0` is not RFC 7233-valid for an empty entity.
-    // Serve it as a plain 200 with an empty body instead of a 206.
-    if file_size == 0 {
-        let response = Response::builder()
-            .status(StatusCode::OK)
-            .header("Content-Type", "application/octet-stream")
-            .header("Content-Length", "0")
-            .body(Body::empty())
-            .expect("build empty GET response");
-        return Ok(response);
-    }
-
-    let range_header = match headers.get("range").or_else(|| headers.get("Range")) {
-        Some(h) => h.to_str().unwrap_or("").to_string(),
-        None => format!("bytes=0-{}", file_size.saturating_sub(1)),
-    };
-
-    let byte_range = match parse_range_header(&range_header, file_size) {
-        Ok(r) => r,
-        Err(e) => {
-            warn!("bad Range header {range_header}: {e}");
-            return Err(StatusCode::BAD_REQUEST);
-        }
-    };
-
-    // An explicit range whose end exceeds the object size (or an open-ended
-    // range whose start is past EOF) is unsatisfiable per RFC 7233 and must be
-    // 416, not 413. This is checked before the 512 MiB span cap so an
-    // out-of-bounds request is never misreported as too large.
-    if byte_range.end > file_size || byte_range.start > file_size {
-        warn!("range {byte_range:?} out of bounds for {key} (file_size {file_size})");
-        return Err(StatusCode::RANGE_NOT_SATISFIABLE);
-    }
-
-    // An open-ended range at exactly EOF (`bytes=<file_size>-`) is a valid
-    // zero-length read. A zero-length entity has no satisfiable byte position to
-    // express in a 206 `Content-Range`, so serve it as a plain 200 with an empty
-    // body, mirroring the zero-byte-object case above, rather than falling
-    // through to `get_range` (which would build a malformed 206 header).
-    if byte_range.start == byte_range.end {
-        let response = Response::builder()
-            .status(StatusCode::OK)
-            .header("Content-Type", "application/octet-stream")
-            .header("Content-Length", "0")
-            .body(Body::empty())
-            .expect("build empty range response");
-        return Ok(response);
-    }
-
-    // The cache assembles the requested span contiguously in memory, so cap the
-    // single-range read at the same limit as the multi-range POST path to bound
-    // peak allocation. The client falls back to the direct store on any non-2xx,
-    // so an oversized read still succeeds (just uncached).
-    if byte_range.end - byte_range.start > MAX_TOTAL_REQUESTED_BYTES {
-        warn!(
-            "rejected range {byte_range:?} for {key}: requested bytes exceed max {MAX_TOTAL_REQUESTED_BYTES}"
-        );
-        return Err(StatusCode::PAYLOAD_TOO_LARGE);
-    }
-
-    match state.cache.get_range(&key, byte_range.clone()).await {
-        Ok(data) => {
-            let content_length = data.len();
-            let response = Response::builder()
-                .status(StatusCode::PARTIAL_CONTENT)
-                .header("Content-Type", "application/octet-stream")
-                .header("Content-Length", content_length.to_string())
-                .header(
-                    "Content-Range",
-                    format!(
-                        "bytes {}-{}/{}",
-                        byte_range.start,
-                        byte_range.end.saturating_sub(1),
-                        file_size
-                    ),
-                )
-                .body(Body::from(data))
-                .expect("build GET response");
-            Ok(response)
-        }
-        Err(e) => {
-            if let Some(RangeError::OutOfBounds { .. }) = e.downcast_ref::<RangeError>() {
-                warn!("range {byte_range:?} out of bounds for {key}: {e}");
-                return Err(StatusCode::RANGE_NOT_SATISFIABLE);
-            }
-            error!("get_range {key} {byte_range:?}: {e:?}");
-            Err(StatusCode::INTERNAL_SERVER_ERROR)
-        }
-    }
-}
-
-/// Maximum number of ranges accepted in a single multi-range request. A
-/// parquet/block reader fetches at most a few thousand column chunks per file,
-/// so this is comfortably above legitimate use while bounding per-request work.
-const MAX_RANGES_PER_REQUEST: usize = 4096;
-
-/// Maximum total requested bytes (summed across all ranges) for a single
-/// multi-range request. The handler assembles all results in memory, so this
-/// caps peak allocation regardless of how many ranges overlap the same bytes.
-const MAX_TOTAL_REQUESTED_BYTES: u64 = 512 * 1024 * 1024; // 512 MiB
-
-#[derive(Deserialize)]
-struct RangesRequest {
-    ranges: Vec<[u64; 2]>,
-}
-
-async fn post_ranges_handler(
-    Path(key): Path<String>,
-    State(state): State<AppState>,
-    body: Bytes,
-) -> Result<Response, StatusCode> {
-    if let Err(e) = validate_key(&key, &state.allowed_prefix) {
-        warn!("rejected key {key}: {e}");
-        return Err(StatusCode::BAD_REQUEST);
-    }
-
-    let req: RangesRequest = match serde_json::from_slice(&body) {
-        Ok(r) => r,
-        Err(e) => {
-            warn!("bad ranges JSON: {e}");
-            return Err(StatusCode::BAD_REQUEST);
-        }
-    };
-
-    // Bound the number of ranges to cap per-request work on this public
-    // authenticated endpoint.
-    if req.ranges.len() > MAX_RANGES_PER_REQUEST {
-        warn!(
-            "rejected {n} ranges for {key}: exceeds max {MAX_RANGES_PER_REQUEST}",
-            n = req.ranges.len()
-        );
-        return Err(StatusCode::BAD_REQUEST);
-    }
-
-    // Reject inverted/degenerate ranges (e.g. `[100, 50]` or `[50, 50]`),
-    // matching the single-range path's `parse_range_header` validation. An
-    // empty or backwards range would otherwise silently yield 0-length data.
-    // While iterating, sum the requested bytes to bound the in-memory assembled
-    // response (overlapping ranges can otherwise amplify allocation).
-    let mut total_requested: u64 = 0;
-    for &[s, e] in &req.ranges {
-        if s >= e {
-            warn!("rejected inverted range [{s}, {e}] for {key}");
-            return Err(StatusCode::BAD_REQUEST);
-        }
-        total_requested = total_requested.saturating_add(e - s);
-        if total_requested > MAX_TOTAL_REQUESTED_BYTES {
-            warn!(
-                "rejected ranges for {key}: total requested bytes exceeds max {MAX_TOTAL_REQUESTED_BYTES}"
-            );
-            return Err(StatusCode::PAYLOAD_TOO_LARGE);
-        }
-    }
-
-    let ranges: Vec<std::ops::Range<u64>> = req.ranges.iter().map(|&[s, e]| s..e).collect();
-
-    match state.cache.get_ranges(&key, &ranges).await {
-        Ok(results) => {
-            let mut buf = BytesMut::new();
-            for chunk in &results {
-                buf.put_u64_le(chunk.len() as u64);
-                buf.put_slice(chunk);
-            }
-            let response = Response::builder()
-                .status(StatusCode::OK)
-                .header("Content-Type", "application/octet-stream")
-                .body(Body::from(buf.freeze()))
-                .expect("build ranges response");
-            Ok(response)
-        }
-        Err(e) => {
-            if let Some(RangeError::OutOfBounds { .. }) = e.downcast_ref::<RangeError>() {
-                warn!("ranges {ranges:?} out of bounds for {key}: {e}");
-                return Err(StatusCode::RANGE_NOT_SATISFIABLE);
-            }
-            if is_not_found(&e) {
-                return Err(StatusCode::NOT_FOUND);
-            }
-            error!("get_ranges {key}: {e:?}");
-            Err(StatusCode::INTERNAL_SERVER_ERROR)
-        }
-    }
-}
 
 #[micromegas_main(interop_max_level = "info")]
 async fn main() -> Result<(), Box<dyn std::error::Error>> {
